distributed-db (3)/distributed-db/requirements.txt
mysql-connector-python==8.2.0
pymysql==1.1.0
pydantic==2.5.0
python-dotenv==1.0.0
